Move `functions` to `routes` and remove useless
