--- bin/resetsql.sh
@@ -1,1 +1,0 @@
-rm ./resources/main.sqlite3 && touch ./resources/main.sqlite3
