+++ routes/series.js
